+++ src/core/src/lib/utils.ts
@@ -4,6 +4,13 @@ import { AbstractControl } from '@angular/forms';
 import { FormlyFieldConfigCache } from './models';
 import { ChangeDetectorRef, ComponentRef, TemplateRef } from '@angular/core';
 
+export function disableTreeValidityCall(form: any, callback: Function) {
+  const _updateTreeValidity = form._updateTreeValidity.bind(form);
+  form._updateTreeValidity = () => {};
+  callback();
+  form._updateTreeValidity = _updateTreeValidity;
+}
+
 export function getFieldId(formId: string, field: FormlyFieldConfig, index: string | number) {
   if (field.id) {
     return field.id;
@@ -12,6 +19,11 @@ export function getFieldId(formId: string, field: FormlyFieldConfig, index: stri
   if (!type && field.template) {
     type = 'template';
   }
+
+  if (isFunction(type)) {
+    type = (type as any).prototype.constructor.name;
+  }
+
   return [formId, type, field.key, index].join('_');
 }